+++ src/App.jsx
@@ -4,8 +4,8 @@ import { getAllData } from './util/index';
 import Handshake from '@mui/icons-material/Handshake';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
-import ListHotels from './components/hotels/ListHotels';
-import HotelDetails from './components/hotels/HotelDetails';
+import HotelsPage from './pages/HotelsPage';
+import HotelDetailsPage from './pages/HotelDetailsPage';
 
 const URL = `${import.meta.env.VITE_APP_API_URL}/api/v1`;
 
@@ -40,8 +40,8 @@ const App = () => {
             </Typography>
           }
         />
-        <Route path="hotels" element={<ListHotels />} />
-        <Route path="hotels/:id" element={<HotelDetails />} />
+        <Route path="hotels" element={<HotelsPage />} />
+        <Route path="hotels/:id" element={<HotelDetailsPage />} />
       </Routes>
     </BrowserRouter>
   );
